Split non-training code from NeuralNetwork.h/cc into Perceptron.h/cc
