+++ CMakeLists.txt
@@ -13,9 +13,10 @@ include_directories (${PROJECT_SOURCE_DIR}/src)
 
 set (LIB_SOURCES
   ${PROJECT_SOURCE_DIR}/src/ActivationFunction.cc
-  ${PROJECT_SOURCE_DIR}/src/MultiLayerPerceptron.cc
-  ${PROJECT_SOURCE_DIR}/src/NeuralNetwork.cc
+  ${PROJECT_SOURCE_DIR}/src/FeedForwardNeuralNetwork.cc
+  ${PROJECT_SOURCE_DIR}/src/MultiLayerNeuralTopology.cc
   ${PROJECT_SOURCE_DIR}/src/NeuronContainer.cc
+  ${PROJECT_SOURCE_DIR}/src/Perceptron.cc
   ${PROJECT_SOURCE_DIR}/src/RandomWrapper.cc
   ${PROJECT_SOURCE_DIR}/src/RecurrentNeuralNetwork.cc
   ${PROJECT_SOURCE_DIR}/src/TrainingData.cc)
